+++ src/main/java/io/dataspaceconnector/controller/message/ids/SubscriptionMessageController.java
@@ -12,6 +12,10 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
+ *  Contributors:
+ *       sovity GmbH
+ *
  */
 package io.dataspaceconnector.controller.message.ids;
 
@@ -27,6 +31,7 @@ import io.dataspaceconnector.controller.message.tag.MessageName;
 import io.dataspaceconnector.controller.util.ResponseCode;
 import io.dataspaceconnector.controller.util.ResponseDescription;
 import io.dataspaceconnector.controller.util.ResponseUtils;
+import io.dataspaceconnector.extension.telemetry.TelemetrySpan;
 import io.dataspaceconnector.model.subscription.SubscriptionDesc;
 import io.dataspaceconnector.service.message.builder.type.SubscriptionRequestService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -98,6 +103,7 @@ public class SubscriptionMessageController {
             @ApiResponse(responseCode = "502", description = "Bad gateway")})
     @PreAuthorize("hasPermission(#recipient, 'rw')")
     @ResponseBody
+    @TelemetrySpan(name = "POST /api/ids/subscribe")
     public ResponseEntity<Object> subscribe(
             @Parameter(description = "The recipient url.", required = true)
             @RequestParam("recipient") final URI recipient,
@@ -148,6 +154,7 @@ public class SubscriptionMessageController {
             @ApiResponse(responseCode = "500", description = "Internal server error")})
     @PreAuthorize("hasPermission(#recipient, 'rw')")
     @ResponseBody
+    @TelemetrySpan(name = "POST /api/ids/unsubscribe")
     public ResponseEntity<Object> unsubscribe(
             @Parameter(description = "The recipient url.", required = true)
             @RequestParam("recipient") final URI recipient,