+++ src/main/java/io/dataspaceconnector/controller/routing/RoutesController.java
@@ -12,6 +12,10 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
+ *  Contributors:
+ *       sovity GmbH
+ *
  */
 package io.dataspaceconnector.controller.routing;
 
@@ -21,6 +25,7 @@ import io.dataspaceconnector.controller.routing.tag.CamelDescription;
 import io.dataspaceconnector.controller.routing.tag.CamelName;
 import io.dataspaceconnector.controller.util.ResponseCode;
 import io.dataspaceconnector.controller.util.ResponseDescription;
+import io.dataspaceconnector.extension.telemetry.TelemetrySpan;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
@@ -80,6 +85,7 @@ public class RoutesController {
                     description = ResponseDescription.BAD_REQUEST),
             @ApiResponse(responseCode = ResponseCode.INTERNAL_SERVER_ERROR,
                     description = ResponseDescription.INTERNAL_SERVER_ERROR)})
+    @TelemetrySpan(name = "POST /api/camel/routes")
     public ResponseEntity<Object> addRoutes(@RequestParam("file") final MultipartFile file) {
         try {
             if (file == null) {
@@ -133,6 +139,7 @@ public class RoutesController {
             @ApiResponse(responseCode = ResponseCode.OK, description = ResponseDescription.OK),
             @ApiResponse(responseCode = ResponseCode.INTERNAL_SERVER_ERROR,
                     description = ResponseDescription.INTERNAL_SERVER_ERROR)})
+    @TelemetrySpan(name = "DELETE /api/camel/routes/{routeId}")
     public ResponseEntity<Object> removeRoute(@PathVariable("routeId") final String routeId) {
         try {
             camelContext.getRouteController().stopRoute(routeId);